pyproject.toml
[tool.poetry]
name = "dp2rathena"
version = "0.4.0"
description = "Convert Divine-Pride API data to rAthena YAML"
authors = ["Eric Liu <latiosworks@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Latiosu/dp2rathena"
documentation = "https://github.com/Latiosu/dp2rathena"
keywords = ["api", "rathena"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = ">=3.6"
python-dotenv = "^0.15.0"
PyYAML = "^5.3.1"
tortilla = "^0.5.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
tox = "^3.20.1"
tox-travis = "^0.12"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
dp2rathena = "dp2rathena.cli:dp2rathena"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
